+++ routes/comment.js
@@ -3,8 +3,8 @@ const express = require('express');
 const router = express.Router();
 
 const { verify } = require('../middlewares/verifyToken');
+const { isAuthorized } = require('../middlewares/isAuthorizedComment');
 const commentCtrl = require('../controllers/comments');
-const { isAuthorized } = require('../middlewares/isAuthorized');
 
 router.delete('/:id', verify, isAuthorized, commentCtrl.deleteCommentById);
